fixed pylint and pytest yml
