+++ backend/.pylintrc
@@ -1,2 +1,2 @@
 [MASTER]
-disable=broad-exception-caught,too-many-instance-attributes,too-many-positional-arguments,too-many-arguments,too-many-return-statements,no-member
+disable=broad-exception-caught,too-many-instance-attributes,too-many-positional-arguments,too-many-arguments,too-many-return-statements,no-member,too-many-locals,line-too-long
